Move example utils to own file
